put bad channels into json file, script in experiment folder rather than tests, request of OW
