+++ experiment/bad_ch_annotations.json
@@ -1,0 +1,282 @@
+{
+    "1d547041-230a-4af3-ba6a-7287de2bdec3": {
+        "idead": [
+            191
+        ],
+        "inoisy": [],
+        "ioutside": []
+    },
+    "4cb60c5c-d15b-4abd-8cfd-776bc5a81dbe": {
+        "idead": [
+            16,
+            29,
+            191,
+            357
+        ],
+        "inoisy": [],
+        "ioutside": [
+            358,
+            359,
+            360,
+            361,
+            362,
+            363,
+            364,
+            365,
+            366,
+            367,
+            368,
+            369,
+            370,
+            371,
+            372,
+            373,
+            374,
+            375,
+            376,
+            377,
+            378,
+            379,
+            380,
+            381,
+            382,
+            383
+        ]
+    },
+    "ce0dc660-f19e-46a3-94f9-646bebae6805": {
+        "idead": [
+            29,
+            36,
+            39,
+            40,
+            191
+        ],
+        "inoisy": [
+            133,
+            235
+        ],
+        "ioutside": [
+            274,
+            275,
+            276,
+            277,
+            278,
+            279,
+            280,
+            281,
+            282,
+            283,
+            284,
+            285,
+            286,
+            287,
+            288,
+            289,
+            290,
+            291,
+            292,
+            293,
+            294,
+            295,
+            296,
+            297,
+            298,
+            299,
+            300,
+            301,
+            302,
+            303,
+            304,
+            305,
+            306,
+            307,
+            308,
+            309,
+            310,
+            311,
+            312,
+            313,
+            314,
+            315,
+            316,
+            317,
+            318,
+            319,
+            320,
+            321,
+            322,
+            323,
+            324,
+            325,
+            326,
+            327,
+            328,
+            329,
+            330,
+            331,
+            332,
+            333,
+            334,
+            335,
+            336,
+            337,
+            338,
+            339,
+            340,
+            341,
+            342,
+            343,
+            344,
+            345,
+            346,
+            347,
+            348,
+            349,
+            350,
+            351,
+            352,
+            353,
+            354,
+            355,
+            356,
+            357,
+            358,
+            359,
+            360,
+            361,
+            362,
+            363,
+            364,
+            365,
+            366,
+            367,
+            368,
+            369,
+            370,
+            371,
+            372,
+            373,
+            374,
+            375,
+            376,
+            377,
+            378,
+            379,
+            380,
+            381,
+            382,
+            383
+        ]
+    },
+    "3b729602-20d5-4be8-a10e-24bde8fc3092": {
+        "idead": [
+            191
+        ],
+        "inoisy": [],
+        "ioutside": [
+            356,
+            357,
+            358,
+            359,
+            360,
+            361,
+            362,
+            363,
+            364,
+            365,
+            366,
+            367,
+            368,
+            369,
+            370,
+            371,
+            372,
+            373,
+            374,
+            375,
+            376,
+            377,
+            378,
+            379,
+            380,
+            381,
+            382,
+            383
+        ]
+    },
+    "a9c9df46-85f3-46ad-848d-c6b8da4ae67c": {
+        "idead": [
+            191
+        ],
+        "inoisy": [],
+        "ioutside": [
+            348,
+            349,
+            350,
+            351,
+            352,
+            353,
+            354,
+            355,
+            356,
+            357,
+            358,
+            359,
+            360,
+            361,
+            362,
+            363,
+            364,
+            365,
+            366,
+            367,
+            368,
+            369,
+            370,
+            371,
+            372,
+            373,
+            374,
+            375,
+            376,
+            377,
+            378,
+            379,
+            380,
+            381,
+            382,
+            383
+        ]
+    },
+    "03d2d8d1-a116-4763-8425-4ef7b1c1bd35": {
+        "idead": [
+            131,
+            191,
+            235
+        ],
+        "inoisy": [],
+        "ioutside": []
+    },
+    "fe380793-8035-414e-b000-09bfe5ece92a": {
+        "idead": [
+            131,
+            191
+        ],
+        "inoisy": [],
+        "ioutside": []
+    },
+    "19c9caea-2df8-4097-92f8-0a2bad055948": {
+        "idead": [],
+        "inoisy": [
+            191
+        ],
+        "ioutside": []
+    },
+    "25a9182c-4795-4768-af47-98975d2d2a8a": {
+        "idead": [],
+        "inoisy": [
+            191
+        ],
+        "ioutside": []
+    }
+}
